API_services/DOWNLOAD_PHOTOS_FROM_NASA_SpaceX/requirements.txt
python-dotenv==0.21.0
requests==2.28.2
python-telegram-bot==13.15
